Docs

+++ malcolm/parts/demo/counterpart.py
@@ -7,7 +7,7 @@ class CounterPart(Part):
     counter = None
 
     def create_attributes(self):
-        self.counter = NumberMeta("uint32", "A counter").make_attribute()
+        self.counter = NumberMeta("float64", "A counter").make_attribute()
         yield "counter", self.counter, self.counter.set_value
 
     @method_takes()
